requirements.txt
python-telegram-bot==20.7
openai==1.3.7
requests==2.31.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
Pillow==10.1.0
pytesseract==0.3.10
schedule==1.2.0
pytz==2023.3
aiohttp==3.9.1
asyncio==3.4.3
